Issue: #21 Updated the style of the page to match the rest of the website and look more appealing to a user
The calender page now looks nice and similar to the rest of the website. It matches the website style we decided on. In the future more functionality will be added to this page in order to allow the user to see what tasks need to be completed on which days in order to meet their goals