Issue: #27 Improved the UI and fucntionality of adding and removing tasks to the calender
On each day when adding a task it small bubble appear around the text as a seperator between other tasks on the same day. I also added the feature of removing tasks. When scrolling over a task it will show a little X button all the way to the right which you can click to remove the task. The numbering for each day has also been moved to the top left of each days cell and out of the way of tasks.