Issue: #29 Created Sign Up and Login pages
When signing up the user is directed to the login page and after loging in the user is gratedn access to the website. Both pages check if the username and password fields are checked. The sign up page currently does not add to a database and the login page does not verify with a database either. This fucntionality will be added in the future