Issue: #38 Added inital tasks to show up to mimic final product for MVP
I added intial tasks that are fake just to demonstrate how the page will look for a user. In the finial project the tasks the calender page opens up with will be pulled from a database, but as of now it is not but is in place just to show how it will look for the user.