Issue #41 Creates web framework to run website on through node/nodemon
I set up the web framework for a test webpage in order to see how to run a website on a server. I used node to run the website on the server but nodemon for development. Nodemon updates the website server live which allows for convenient web development. Node severs require you to shut down the server and start it up again to keep up with updates.

Because I set up a test web framework, it should be quite simple to start integrating our webpages by simplying adding it to the public file.